+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the enable authenticator endpoint for the V2 API's. CNJR-9135
 - Dynamic secrets are now supported in the batch secret retrieval API. CNJR-9172
 - Added the delete authenticators endpoint for the V2 API's. CNJR-9134
+- Added warning for annotation keys matching known policy attribute names. CNJR-9836
 
 ### Fixed
 - Attempt to load a policy that references a non-existent resource now